+++ mark1/src/main/java/com/starkindustries/dao/ShippingMapper.java
@@ -3,13 +3,13 @@ package com.starkindustries.dao;
 import com.starkindustries.pojo.Shipping;
 
 public interface ShippingMapper {
-    int deleteByPrimaryKey(Byte id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(Shipping record);
 
     int insertSelective(Shipping record);
 
-    Shipping selectByPrimaryKey(Byte id);
+    Shipping selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(Shipping record);
